Removing a uint that was causing problems with build
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@3691 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/analysis/current/parallelanalysis/DuchampParallel.cc
@@ -1967,7 +1967,7 @@ namespace askap {
 	IPosition shape = imagePtr->shape();
 	long *dim = new long[shape.size()];
 
-	for (uint i = 0; i < shape.size(); i++) {
+	for (size_t i = 0; i < shape.size(); i++) {
 	  dim[i] = shape(i);
 	  ASKAPCHECK(dim[i] > 0, "Negative dimension: dim[" << i << "]=" << dim[i]);
 	}
